Crr and CdA Regression Explainer
An explainer from Robert Chung for a regression method to estimate CdA and Crr from field data.

The key point being the generation of additional laps from a multi-lap run -- e.g. 3 x 18s laps of a velodrome
would normally be considered as 3 laps to fit against. Robert's approach is to create 36 laps from this data (!).
The first lap starts at 0s and lasts for 18s, the second lap at 1s ... up to the 36th lap starting at 37s.

We would then be able to get 36 estimates of Crr/CdA and calculate a mean and confidence interval (!).

The document committed explains how that regression would work mathematically, and how we
would adapt for working in the field (we need really good altitude data).